moved boilerplate codepen code into ExampleCard, added documentation
