+++ packages/office-ui-fabric-react/src/components/Label/Label.doc.tsx
@@ -1,6 +1,5 @@
 import * as React from 'react';
 import { LabelBasicExample } from './examples/Label.Basic.Example';
-import { LabelCodepenButton } from './LabelCodepenButton';
 
 import { IDocPageProps } from '../../common/DocPage.types';
 import { LabelStatus } from './Label.checklist';
@@ -18,7 +17,7 @@ export const LabelPageProps: IDocPageProps = {
       title: 'Label',
       code: LabelBasicExampleCode,
       view: <LabelBasicExample />,
-      codepen: LabelCodepenButton()
+      codepen: require('!raw-loader!office-ui-fabric-react/src/components/Label/codepen_examples/Label.Basic.CodepenExample.txt') as string
     }
   ],
   propertiesTablesSources: [require<string>('!raw-loader!office-ui-fabric-react/src/components/Label/Label.types.ts')],
